Move shared components to util
