+++ api/web/src/components/util/Coverage.vue
